Longest Monotonic Sequence Problem
There are two parts of the project.
The purpose of the part one is to find the monotonic subsequence of the
given input array, whose total weight is maximal.
The output consist of  3 lines.  The first line has the total weight.
The second line has indexes of the optimal subsequence, and the third
line has the actual subsequence itself.
The purpose of the part two is to find a subsequence, that violates
monotonicity at most k times, such that the sum of the weights, minus
the penalties, is a maximum.
The penalty for a violation (a, b) with a>b, is equal to the
weight of b, times (a-b).
Similar to part one, the output of the part two also consist of 3 lines.
The first line has the total weight. The second line has indexes of the
optimal subsequence, and the third line has the actual subsequence
itself.